# jwt: reduce heap allocations in jwt.Parse