--- a/jwt/token.go
+++ b/jwt/token.go
@@ -16,7 +16,7 @@ type Token struct {
 }
 
 func Parse(value string) (*Token, error) {
-	parts := strings.Split(value, ".")
+	parts := strings.SplitN(value, ".", 4)
 	if len(parts) != 3 {
 		return nil, errInvalidToken
 	}